check_package location moved to base
